requirements.txt
allure-pytest==2.5.0
pytest==3.7.1
